hotfix: rating

+++ biz/src/main/kotlin/xcode/biz/domain/response/product/ProductListResponse.kt
@@ -9,7 +9,7 @@ class ProductListResponse {
     var seat: Int? = 0
     var name = ""
     var price: Int? = null
-    var rating: Int? = null
+    var rating: Float? = null
     var deliverable = false
     var brand: VehicleBrand? = null
 }
